Moved EntryType to holochain_dna. Added unit_tests
