+++ core/src/chain/mod.rs
@@ -12,9 +12,10 @@ use chain::{
 use error::HolochainError;
 use hash_table::{
     entry::Entry,
-    sys_entry::{EntryType, ToEntry},
+    sys_entry::ToEntry,
     HashTable,
 };
+use holochain_dna::entry_type::EntryType;
 use json::ToJson;
 use riker::actors::*;
 use serde_json;
