+++ core/src/dht/dht_reducers.rs
@@ -5,10 +5,9 @@ use cas::{content::AddressableContent, storage::ContentAddressableStorage};
 use context::Context;
 use dht::dht_store::DhtStore;
 use eav::EntityAttributeValueStorage;
-use hash_table::{
-    entry::Entry, sys_entry::EntryType,
-};
+use hash_table::entry::Entry;
 use std::sync::Arc;
+use holochain_dna::entry_type::EntryType;
 
 // A function that might return a mutated DhtStore
 type DhtReducer<CAS, EAVS> =
@@ -87,7 +86,7 @@ pub(crate) fn commit_app_entry<CAS, EAVS>(
 {
     // pre-condition: if app entry_type must be valid
     // get entry_type definition
-    let dna = context.state().unwrap().nucleus().dna().expect("Must have DNA to commit an entry.");
+    let dna = context.state().unwrap().nucleus().dna().expect("Must have DNA to commit an app entry.");
     let maybe_def = dna.get_entry_type_def(&entry_type.to_string());
     if maybe_def.is_none() {
         return None;
@@ -102,7 +101,6 @@ pub(crate) fn commit_app_entry<CAS, EAVS>(
         return None;
     }
     // ...and publish to the network if its not private
-    // System entry types are private
     if entry_type_def.sharing.clone().can_publish()  {
         new_store.network_mut().publish(entry);
     }
