+++ core/src/hash_table/sys_entry.rs
@@ -1,100 +1,12 @@
 use cas::content::AddressableContent;
 use hash_table::entry::Entry;
 use holochain_agent::{Agent, Identity};
-use holochain_dna::Dna;
+use holochain_dna::{Dna, entry_type::EntryType};
 use serde_json;
-use std::{
-    fmt::{Display, Formatter, Result as FmtResult},
-    str::FromStr,
-};
 
 pub trait ToEntry {
     fn to_entry(&self) -> (EntryType, Entry);
     fn from_entry(&Entry) -> Self;
-}
-
-//-------------------------------------------------------------------------------------------------
-// Entry Type
-//-------------------------------------------------------------------------------------------------
-
-// Macro for statically concatanating the system entry prefix for entry types of system entries
-macro_rules! sys_prefix {
-    ($s:expr) => {
-        concat!("%", $s)
-    };
-}
-
-
-// Enum for listing all System Entry Types
-// Variant `Data` is for user defined entry types
-#[derive(Debug, Clone, PartialEq, Hash, Serialize, Deserialize)]
-pub enum EntryType {
-    AgentId,
-    Deletion,
-    App(String),
-    Dna,
-    ChainHeader,
-    Key,
-    Link,
-    Migration,
-    /// TODO #339 - This is different kind of SystemEntry for the DHT only.
-    /// Should be moved into a different enum for DHT entry types.
-    LinkList,
-}
-
-impl EntryType {
-    pub fn is_app(self) -> bool {
-        match self  {
-            EntryType::App(_) => true,
-            _ => false,
-        }
-    }
-    pub fn is_sys(self) -> bool { !self.is_app() }
-
-    pub fn has_sys_prefix(entry_type_name: &str) -> bool {
-        &entry_type_name[0..1] == "%"
-    }
-}
-
-impl FromStr for EntryType {
-    type Err = usize;
-    // Note: Function always return Ok()
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            sys_prefix!("agent_id") => Ok(EntryType::AgentId),
-            sys_prefix!("deletion") => Ok(EntryType::Deletion),
-            sys_prefix!("dna") => Ok(EntryType::Dna),
-            sys_prefix!("chain_header") => Ok(EntryType::ChainHeader),
-            sys_prefix!("key") => Ok(EntryType::Key),
-            sys_prefix!("link") => Ok(EntryType::Link),
-            sys_prefix!("link_list") => Ok(EntryType::LinkList),
-            sys_prefix!("migration") => Ok(EntryType::Migration),
-            _ => Ok(EntryType::App(s.to_string())),
-        }
-    }
-}
-
-impl Display for EntryType {
-    fn fmt(&self, f: &mut Formatter) -> FmtResult {
-        write!(f, "{}", self.as_str())
-    }
-}
-
-impl EntryType {
-    pub fn as_str(&self) -> &str {
-        let ret = match *self {
-            EntryType::App(ref s) => s,
-            EntryType::AgentId => sys_prefix!("agent_id"),
-            EntryType::Deletion => sys_prefix!("deletion"),
-            EntryType::Dna => sys_prefix!("dna"),
-            EntryType::ChainHeader => sys_prefix!("chain_header"),
-            EntryType::Key => sys_prefix!("key"),
-            EntryType::Link => sys_prefix!("link"),
-            EntryType::LinkList => sys_prefix!("link_list"),
-            EntryType::Migration => sys_prefix!("migration"),
-        };
-        ret
-    }
 }
 
 //-------------------------------------------------------------------------------------------------
@@ -137,8 +49,8 @@ pub mod tests {
     extern crate test_utils;
 
     use action::{Action, ActionWrapper};
-    use hash_table::sys_entry::{EntryType, ToEntry};
-    use std::str::FromStr;
+    use hash_table::sys_entry::ToEntry;
+    use holochain_dna::entry_type::EntryType;
 
     use cas::content::AddressableContent;
     use instance::{tests::test_context, Instance, Observer};
@@ -167,7 +79,7 @@ pub mod tests {
             .iter()
             .find(|aw| match aw.action() {
                 Action::Commit(entry_type, entry) => {
-                    assert_eq!(entry_type, &EntryType::Dna,);
+                    assert_eq!(entry_type, &EntryType::Dna);
                     assert_eq!(entry.content(), dna_entry.content());
                     true
                 }
@@ -205,26 +117,4 @@ pub mod tests {
                 _ => false,
             });
     }
-
-    #[test]
-    /// converting a str to an EntryType and back
-    fn test_from_as_str() {
-        for (type_str, variant) in vec![
-            (sys_prefix!("agent_id"), EntryType::AgentId),
-            (sys_prefix!("deletion"), EntryType::Deletion),
-            (sys_prefix!("dna"), EntryType::Dna),
-            (sys_prefix!("chain_header"), EntryType::ChainHeader),
-            (sys_prefix!("key"), EntryType::Key),
-            (sys_prefix!("link"), EntryType::Link),
-            (sys_prefix!("migration"), EntryType::Migration),
-        ] {
-            assert_eq!(
-                variant,
-                EntryType::from_str(type_str).expect("could not convert str to EntryType")
-            );
-
-            assert_eq!(type_str, variant.as_str(),);
-        }
-    }
-
 }
