+++ core/src/instance.rs
@@ -276,7 +276,7 @@ pub mod tests {
     use agent::state::ActionResponse;
     use context::Context;
     use futures::executor::block_on;
-    use hash_table::sys_entry::EntryType;
+    use holochain_dna::entry_type::EntryType;
     use holochain_agent::Agent;
     use holochain_dna::{zome::Zome, Dna};
     use logger::Logger;
